Manipulation: privatize buildFragment() function
Fixes gh-2224
