+++ src/manipulation/var/rscriptType.js
@@ -1,0 +1,3 @@
+define(function() {
+	return ( /^$|\/(?:java|ecma)script/i );
+});
