+++ src/manipulation/var/rtagName.js
@@ -1,0 +1,3 @@
+define(function() {
+	return ( /<([\w:-]+)/ );
+});
